salon-flask/requirements.txt
boto3==1.26.146
Flask==2.3.2
matplotlib==3.7.1
numpy==1.24.3
Pillow==9.5.0
gunicorn==20.1.0

